boot: zephyr: defines FLASH device for external NOR
With CONFIG_STM32_MEMMAP flag get NOR flash 32 bytes header
Read the NOR flash to get header of the image.
The FLASH_DEVICE is now given by the parent of the
"st,stm32-xspi-nor" node, which  is the "st,stm32-xspi" compatible
same for qspi/ospi instances of stm32 devices.

Signed-off-by: Francois Ramu <francois.ramu@st.com>