sim: Generate smallest possible oversized image
To generate oversized, the simulator needs to know the maximum image
size. To obtain such size, the size of the TLV area is estimated and
when using ECDSA, the actual size of the TLV area in the generated image
was not always equal to the estimated size. This required to add a bit
more data than what should be necessary when creating oversized images,
to ensure the generated images will actually be oversized in most cases.

Thanks to the previous commit, this is no more necessary and it is now
possible to reliably generate oversized images with the smallest size.

Signed-off-by: Thomas Altenbach <thomas.altenbach@legrand.com>