bootutil: swap-scratch: Fix conflicts after support of devices w/o erase
Before the PR #2206 fixing issues in the swap-scratch logic when the
trailer is larger than the last sector was merged, the PR #2114 was
reverted (see #2211). This changes made by #2211 were then reapplied
(see #2216) after #2206 was merged, leading to some conflicts: a call to
swap_erase_trailer_sectors was moved by #2206 but the old call to that
function was added back by #2216 and a new call to
swap_erase_trailer_sectors was added by #2206 but not converted to a
call to swap_scramble_trailer_sectors by #2216.

Signed-off-by: Thomas Altenbach <thomas.altenbach@legrand.com>