boot: stm32n6: Define specific executable region
On STM32N6, MCUBoot is loaded by the BootROM to axisram2 and,
when running in RAMLOAD configuration, application image is loaded
in axisram1.
To support this, define MULTIPLE_EXECUTABLE_RAM_REGIONS and implement
boot_get_image_exec_ram_info() which provides the right executable
address and size based on axisram1 node definition.

Signed-off-by: Erwan Gouriou <erwan.gouriou@st.com>