bootutil: loader: Expose routine to open/close all flash areas
Opening/closing all the flash areas is already implemented in loader.c
and will be needed in boot_serial to initialize a bootloader state. To
avoid code duplication, this commit creates a routine to open all the
flash areas of a state and expose in bootutil_priv.h that routine and
the existing close_all_flash_areas routine to be able to access them
from boot_serial.

Signed-off-by: Thomas Altenbach <thomas.altenbach@legrand.com>